# openapi-to-postman/doc/openapi.json
{"openapi":"3.1.0","info":{"title":"stock link app","version":"0.1.0"},"paths":{"/api/v1/jpx/stock_info/code/{code}":{"get":{"tags":["jpx"],"summary":"Read Jpx Stock Info Item","description":"Get item by code.","operationId":"jpx-read_jpx_stock_info_item","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string","title":"Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JpxStockInfoPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jpx/stock_info/":{"get":{"tags":["jpx"],"summary":"Read Jpx Stock Info Items","description":"Get all items.","operationId":"jpx-read_jpx_stock_info_items","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JpxStockInfosPublicList"}}}}}}},"/api/v1/jpx/stock_info/tcs":{"get":{"tags":["jpx"],"summary":"Read Jpx Stock Info Items Tcs","description":"Get all items.","operationId":"jpx-read_jpx_stock_info_items_tcs","parameters":[{"name":"industry_17_code","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Industry 17 Code"}},{"name":"industry_33_code","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"integer"}},{"type":"null"}],"title":"Industry 33 Code"}},{"name":"isItems","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Isitems"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JpxStockInfosPublicList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jpx/stock_info/tcs/{market}":{"get":{"tags":["jpx"],"summary":"Read Jpx Stock Info Item Tcs","description":"Get item by market.","operationId":"jpx-read_jpx_stock_info_item_tcs","parameters":[{"name":"market","in":"path","required":true,"schema":{"type":"string","title":"Market"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JpxStockInfosPublicList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jpx/stock_info/industries/{type}":{"get":{"tags":["jpx"],"summary":"Read Jpx Stock Info Industry Names","description":"Get all industries.","operationId":"jpx-read_jpx_stock_info_industry_names","parameters":[{"name":"type","in":"path","required":true,"schema":{"type":"integer","title":"Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndustriesList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jpx/stock_info/industries":{"get":{"tags":["jpx"],"summary":"Read Select Industries","description":"Get all industries.","operationId":"jpx-read_select_industries","parameters":[{"name":"industry_17_code","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Industry 17 Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndustriesList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ix/summary/menu_labels/{head_item_key}":{"get":{"tags":["summary"],"summary":"メニューラベルを取得","description":"## メニューラベルを取得するエンドポイント\n- **機能**: HeadItemKeyからメニューラベル一覧を取得します。\n- **認証不要**\n- **レスポンス形式**: JSON\n- **param1**: head_item_key: str 必須項目","operationId":"summary-read_menu_labels","parameters":[{"name":"head_item_key","in":"path","required":true,"schema":{"type":"string","title":"Head Item Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuLabelList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ix/summary/tree_items/{head_item_key}":{"get":{"tags":["summary"],"summary":"表示リンクアイテムを取得","description":"## 表示リンクアイテムを取得するエンドポイント\n- **機能**: HeadItemKeyから表示リンクのツリーアイテム一覧を取得します。\n- **認証不要**\n- **レスポンス形式**: JSON\n- **param1**: head_item_key: str 必須項目\n- **param2**: attr_value: str 任意項目\n- **param3**: has_children: bool 任意項目\n- **param4**: xlink_arcrole: str 任意項目\n- **param5**: xbrl_type: str 任意項目","operationId":"summary-read_tree_items","parameters":[{"name":"head_item_key","in":"path","required":true,"schema":{"type":"string","title":"Head Item Key"}},{"name":"attr_value","in":"query","required":false,"schema":{"type":"string","title":"Attr Value"}},{"name":"has_children","in":"query","required":false,"schema":{"type":"boolean","title":"Has Children"}},{"name":"xlink_arcrole","in":"query","required":false,"schema":{"type":"string","title":"Xlink Arcrole"}},{"name":"xbrl_type","in":"query","required":false,"schema":{"type":"string","default":"sm","title":"Xbrl Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TreeItemsList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ix/summary/tree_items/context_list/{head_item_key}":{"get":{"tags":["summary"],"summary":"コンテキストリストを取得","description":"## コンテキストリストを取得するエンドポイント\n- **機能**: HeadItemKeyからコンテキストリストを取得します。\n- **認証不要**\n- **レスポンス形式**: JSON\n- **param1**: head_item_key: str 必須項目","operationId":"summary-read_context_list","parameters":[{"name":"head_item_key","in":"path","required":true,"schema":{"type":"string","title":"Head Item Key"}},{"name":"attr_value","in":"query","required":false,"schema":{"type":"string","title":"Attr Value"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Summary-Read Context List"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ix/summary/tree_items/names/{head_item_key}":{"get":{"tags":["summary"],"summary":"名前リストを取得","description":"## 名前リストを取得するエンドポイント\n- **機能**: HeadItemKeyから名前リストを取得します。\n- **認証不要**\n- **レスポンス形式**: JSON\n- **param1**: head_item_key: str 必須項目","operationId":"summary-read_names","parameters":[{"name":"head_item_key","in":"path","required":true,"schema":{"type":"string","title":"Head Item Key"}},{"name":"attr_value","in":"query","required":false,"schema":{"type":"string","title":"Attr Value"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Summary-Read Names"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IndustriesList":{"properties":{"data":{"items":{"$ref":"#/components/schemas/Industry"},"type":"array","title":"Data"}},"type":"object","required":["data"],"title":"IndustriesList"},"Industry":{"properties":{"code":{"type":"integer","title":"Code"},"name":{"type":"string","title":"Name"}},"type":"object","required":["code","name"],"title":"Industry"},"JpxStockInfoPublic":{"properties":{"input_date":{"type":"string","title":"Input Date"},"code":{"type":"string","title":"Code"},"name":{"type":"string","title":"Name"},"market_or_type":{"type":"string","title":"Market Or Type"},"industry_33_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Industry 33 Code"},"industry_33_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry 33 Name"},"industry_17_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Industry 17 Code"},"industry_17_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry 17 Name"},"scale_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Scale Code"},"scale_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scale Name"}},"type":"object","required":["input_date","code","name","market_or_type","industry_33_code","industry_33_name","industry_17_code","industry_17_name","scale_code","scale_name"],"title":"JpxStockInfoPublic"},"JpxStockInfosPublicList":{"properties":{"count":{"type":"integer","title":"Count"},"data":{"items":{"$ref":"#/components/schemas/JpxStockInfoPublic"},"type":"array","title":"Data"}},"type":"object","required":["count","data"],"title":"JpxStockInfosPublicList"},"MenuLabel":{"properties":{"xlink_from":{"type":"string","title":"Xlink From"},"label":{"type":"string","title":"Label"}},"type":"object","required":["xlink_from","label"],"title":"MenuLabel","description":"メニューラベルを表すクラス"},"MenuLabelList":{"properties":{"data":{"items":{"$ref":"#/components/schemas/MenuLabel"},"type":"array","title":"Data"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["data","count"],"title":"MenuLabelList","description":"メニューラベルのリストを表すクラス"},"TreeItem":{"properties":{"id":{"type":"integer","title":"Id"},"xlink_from":{"type":"string","title":"Xlink From"},"xlink_to":{"type":"string","title":"Xlink To"},"attr_value":{"type":"string","title":"Attr Value"},"xlink_arcrole":{"type":"string","title":"Xlink Arcrole"},"xlink_order":{"type":"number","title":"Xlink Order"},"level":{"type":"integer","title":"Level"},"has_children":{"type":"boolean","title":"Has Children"}},"type":"object","required":["id","xlink_from","xlink_to","attr_value","xlink_arcrole","xlink_order","level","has_children"],"title":"TreeItem","description":"ツリーアイテムを表すクラス"},"TreeItemsList":{"properties":{"count":{"type":"integer","title":"Count"},"data":{"items":{"$ref":"#/components/schemas/TreeItem"},"type":"array","title":"Data"}},"type":"object","required":["count","data"],"title":"TreeItemsList","description":"ツリーアイテムのリストを表すクラス"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}